Add files via upload
Polymerase random walk code but with a different way to calculate whether or not the walker has hit the target. First it generates a series of points on the interior of the circle from the origin and then the random walker position is is checked to see if its a distance from any of those, more details in the code comments.