Add files via upload
boundary is now "wrap-around", if it hits the boundary on one side it moves to the other side and then takes a random step. it's possible it might need to be changed to deliberately move away from the boundary and towards the origin if you skating along or repeat hitting of the boundary is seen, but so far it hasn't.